requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest-playwright==0.7.0
requests==2.32.3
python-dotenv==1.0.1
openai==1.66.3
jsonschema==4.23.0